Fix Comment component into CommentElement component, add data/comments.js
