+++ src/data/comments.js
@@ -1,0 +1,18 @@
+const comments = [
+  {
+    id: 1,
+    content: "멋사 11기 화이팅!!",
+    like_users: [2],
+    author: 1,
+    created_at: "2023-02-12T08:09:59.900509Z",
+  },
+  {
+    id: 2,
+    content: "멋사 좋아요~~",
+    like_users: [],
+    author: 2,
+    created_at: "2022-11-22T08:09:54.233976Z",
+  },
+];
+
+export default comments;
