+++ src/data/posts.js
@@ -1,61 +1,61 @@
 const posts = [
   {
-    id: 0,
+    id: 1,
     title: "#1 post",
     content: "My Number one post content",
-    author: { id: 0, username: "지현" },
+    author: { id: 1, username: "지현" },
     tags: [
-      { id: 0, content: "멋사" },
-      { id: 1, content: "개발자" },
-      { id: 2, content: "서울대" },
+      { id: 1, content: "멋사" },
+      { id: 2, content: "개발자" },
+      { id: 3, content: "서울대" },
     ],
     like_users: [1, 2],
     created_at: "2023-02-04T07:42:50.658501Z",
   },
   {
-    id: 1,
+    id: 2,
     title: "#2 post",
     content: "My Number two post content",
-    author: { id: 1, username: "란경" },
+    author: { id: 2, username: "란경" },
     tags: [
-      { id: 0, content: "멋사" },
-      { id: 1, content: "낭만걸" },
+      { id: 1, content: "멋사" },
+      { id: 4, content: "낭만걸" },
     ],
-    like_users: [0],
+    like_users: [4],
     created_at: "2023-02-04T07:42:50.658501Z",
   },
   {
-    id: 2,
+    id: 3,
     title: "#3 post",
     content: "My Number three post content",
-    author: { id: 2, username: "진모" },
+    author: { id: 3, username: "진모" },
     tags: [
-      { id: 1, content: "가모" },
-      { id: 2, content: "등산" },
+      { id: 5, content: "가모" },
+      { id: 6, content: "등산" },
     ],
     like_users: [],
     created_at: "2023-02-04T07:42:50.658501Z",
   },
   {
-    id: 3,
+    id: 4,
     title: "#4 post",
     content: "My Number four post content",
-    author: { id: 3, username: "동민" },
+    author: { id: 4, username: "동민" },
     tags: [
-      { id: 1, content: "천재수교과" },
-      { id: 2, content: "다래끼" },
+      { id: 7, content: "천재수교과" },
+      { id: 8, content: "다래끼" },
     ],
     like_users: [0, 1, 2],
     created_at: "2023-02-04T07:42:50.658501Z",
   },
   {
-    id: 4,
+    id: 5,
     title: "#5 post",
     content: "My Number five post content",
-    author: { id: 4, username: "수혁" },
+    author: { id: 5, username: "수혁" },
     tags: [
-      { id: 0, content: "멋사" },
-      { id: 2, content: "연예인" },
+      { id: 1, content: "멋사" },
+      { id: 9, content: "연예인" },
     ],
     like_users: [3],
     created_at: "2023-02-04T07:42:50.658501Z",
